+++ contracts/interfaces/INursePart.sol
@@ -4,13 +4,22 @@ pragma solidity ^0.8.5;
 import "@openzeppelin/contracts/token/ERC1155/IERC1155.sol";
 
 interface INursePart is IERC1155 {
-    
-    event Mint(address indexed to, uint indexed id, uint amount);
-    event Burn(address indexed from, uint indexed id, uint amount);
-    
+    event Mint(address indexed to, uint256 indexed id, uint256 amount);
+    event Burn(address indexed from, uint256 indexed id, uint256 amount);
+
     function nurseRaid() external view returns (address);
+
     function cloneNurse() external view returns (address);
-	
-    function mint(address to, uint id, uint amount) external;
-    function burn(address from, uint id, uint amount) external;
+
+    function mint(
+        address to,
+        uint256 id,
+        uint256 amount
+    ) external;
+
+    function burn(
+        address from,
+        uint256 id,
+        uint256 amount
+    ) external;
 }
