+++ contracts/interfaces/IRatio.sol
@@ -2,18 +2,21 @@
 pragma solidity ^0.8.5;
 
 interface IRatio {
-    
-    event ChangeLPTokenToMaidPower(uint value);
-    event ChangeMaidPowerToRaidReducedBlock(uint value);
-    event ChangeLPTokenToNursePower(uint value);
-    event ChangeNursePowerToRewardAmount(uint value);
-    event ChangeCorpRewardToNurseReward(uint value);
-    
-    function precision() external view returns (uint);
-    
-    function lpTokenToMaidPower() external view returns (uint);
-    function maidPowerToRaidReducedBlock() external view returns (uint);
-    function lpTokenToNursePower() external view returns (uint);
-    function nursePowerToRewardAmount() external view returns (uint);
-    function corpRewardToNurseReward() external view returns (uint);
+    event ChangeLPTokenToMaidPower(uint256 value);
+    event ChangeMaidPowerToRaidReducedBlock(uint256 value);
+    event ChangeLPTokenToNursePower(uint256 value);
+    event ChangeNursePowerToRewardAmount(uint256 value);
+    event ChangeCorpRewardToNurseReward(uint256 value);
+
+    function precision() external view returns (uint256);
+
+    function lpTokenToMaidPower() external view returns (uint256);
+
+    function maidPowerToRaidReducedBlock() external view returns (uint256);
+
+    function lpTokenToNursePower() external view returns (uint256);
+
+    function nursePowerToRewardAmount() external view returns (uint256);
+
+    function corpRewardToNurseReward() external view returns (uint256);
 }
